+++ examples/basic/index.js
@@ -1,9 +1,10 @@
 import { App } from '@tinyhttp/app'
-import logger from '@tinyhttp/logger'
+import { logger } from '@tinyhttp/logger'
 
 const app = new App()
 
 app
+  .use(logger())
   .get('/', (_, res) => void res.send('<h1>Hello World</h1>'))
   .get('/page/:page/', (req, res) => {
     res.status(200).send(`
@@ -14,5 +15,5 @@ app
     ${JSON.stringify(req.params, null, 2)}
   `)
   })
-  .use(logger())
+
   .listen(3000)