+++ examples/mongodb/index.js
@@ -1,30 +1,30 @@
+// @ts-nocheck
 import { App } from '@tinyhttp/app'
 import dotenv from 'dotenv'
 import mongodb from 'mongodb'
 import assert from 'assert'
+import * as parser from 'body-parsec'
 
 dotenv.config()
 
 const app = new App()
+let db
 
 // connect to mongodb
-const client = mongodb.MongoClient(process.env.DB_URI, {
-  useUnifiedTopology: true
+const client = new mongodb.MongoClient(process.env.DB_URI, {
+  useUnifiedTopology: true,
 })
 const dbName = 'notes'
-client.connect(err => {
+client.connect((err) => {
   assert.equal(null, err)
   console.log('successfully connected with mongodb')
+  db = client.db(dbName)
 })
 
 // get all notes
-app.get('/notes', async (req, res, next) => {
+app.get('/notes', async (_, res, next) => {
   try {
-    const db = client.db(dbName)
-    const r = await db
-      .collection('notes')
-      .find({})
-      .toArray()
+    const r = await db.collection('notes').find({}).toArray()
     res.send(r)
   } catch (err) {
     next(err)
@@ -33,11 +33,11 @@ app.get('/notes', async (req, res, next) => {
 
 // add new note
 app.post('/notes', async (req, res, next) => {
+  await parser.form()(req, res)
   try {
-    const db = client.db(dbName)
-    const r = await db.collection('notes').insertOne({ title: req.query.title, desc: req.query.desc })
+    const r = await db.collection('notes').insertOne({ title: req.body.title, desc: req.body.desc })
     assert.equal(1, r.insertedCount)
-    res.send(`Note with title of "${req.query.title}" has been added`)
+    res.send(`Note with title of "${req.body.title}" has been added`)
   } catch (err) {
     next(err)
   }
@@ -46,8 +46,7 @@ app.post('/notes', async (req, res, next) => {
 // delete note
 app.delete('/notes', async (req, res, next) => {
   try {
-    const db = client.db(dbName)
-    const r = await db.collection('notes').deleteOne({ _id: mongodb.ObjectId(req.query.id) })
+    const r = await db.collection('notes').deleteOne({ _id: new mongodb.ObjectId(req.query.id) })
     assert.equal(1, r.deletedCount)
     res.send(`Note with id of ${req.query.id} has been deleted`)
   } catch (err) {
@@ -58,10 +57,9 @@ app.delete('/notes', async (req, res, next) => {
 // update existing note
 app.put('/notes', async (req, res, next) => {
   try {
-    const db = client.db(dbName)
-    const r = await db
+    await db
       .collection('notes')
-      .findOneAndUpdate({ _id: mongodb.ObjectId(req.query.id) }, { $set: { title: req.query.title, desc: req.query.desc } }, { returnOriginal: false, upsert: true })
+      .findOneAndUpdate({ _id: new mongodb.ObjectId(req.query.id) }, { $set: { title: req.query.title, desc: req.query.desc } }, { returnOriginal: false, upsert: true })
     res.send(`Note with title of ${req.query.title} has been updated`)
   } catch (err) {
     next(err)