+++ packages/app/src/router.ts
@@ -4,21 +4,13 @@ import { Response } from './response'
 
 export type NextFunction = (err?: any) => void | undefined
 
-export type SyncHandler = (
-  req: Request,
-  res: Response,
-  next?: NextFunction
-) => void
+export type SyncHandler = (req: Request, res: Response, next?: NextFunction) => void
 
-export type AsyncHandler = (
-  req: Request,
-  res: Response,
-  next?: NextFunction
-) => Promise<void>
+export type AsyncHandler = (req: Request, res: Response, next?: NextFunction) => Promise<void>
 
 export type Handler = AsyncHandler | SyncHandler
 
-export type ErrorHandler = (err: any, req: Request, res: Response) => void
+export type ErrorHandler = (err: any, req: Request, res: Response, next?: NextFunction) => void
 
 type Method =
   | 'GET'
@@ -28,6 +20,33 @@ type Method =
   | 'HEAD'
   | 'OPTIONS'
   | 'DELETE'
+  | 'CHECKOUT'
+  | 'COPY'
+  | 'LOCK'
+  | 'UNLOCK'
+  | 'MERGE'
+  | 'MKACTIVITY'
+  | 'MKCOL'
+  | 'MOVE'
+  | 'SEARCH'
+  | 'M-SEARCH'
+  | 'NOTIFY'
+  | 'PURGE'
+  | 'REPORT'
+  | 'SUBSCRIBE'
+  | 'UNSUBSCRIBE'
+  | 'TRACE'
+  | 'ACL'
+  | 'CONNECT'
+  | 'BIND'
+  | 'UNBIND'
+  | 'REBIND'
+  | 'LINK'
+  | 'UNLINK'
+  | 'MKCALENDAR'
+  | 'PROPFIND'
+  | 'PROPPATCH'
+  | 'SOURCE'
   | string
 
 type MiddlewareType = 'mw' | 'route'
@@ -76,7 +95,9 @@ const pushMiddleware = (mw: Middleware[]) => ({
     mw.push({ ...mdw, type })
   }
 }
-
+/**
+ * tinyhttp Router. Manages middleware and has HTTP methods aliases, e.g. `app.get`, `app.put`
+ */
 export class Router {
   middleware: Middleware[]
 
@@ -151,6 +172,276 @@ export class Router {
     })
     return this
   }
+  checkout(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'CHECKOUT',
+      type: 'route',
+    })
+    return this
+  }
+  copy(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'COPY',
+      type: 'route',
+    })
+    return this
+  }
+  lock(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'LOCK',
+      type: 'route',
+    })
+    return this
+  }
+  unlock(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'LOCK',
+      type: 'route',
+    })
+    return this
+  }
+  merge(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'MERGE',
+      type: 'route',
+    })
+    return this
+  }
+  mkactivity(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'MKACTIVITY',
+      type: 'route',
+    })
+    return this
+  }
+  mkcol(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'MKCOL',
+      type: 'route',
+    })
+    return this
+  }
+  move(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'MOVE',
+      type: 'route',
+    })
+    return this
+  }
+  search(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'SEARCH',
+      type: 'route',
+    })
+    return this
+  }
+  msearch(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'M-SEARCH',
+      type: 'route',
+    })
+    return this
+  }
+  notify(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'NOTIFY',
+      type: 'route',
+    })
+    return this
+  }
+  purge(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'PURGE',
+      type: 'route',
+    })
+    return this
+  }
+  report(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'REPORT',
+      type: 'route',
+    })
+    return this
+  }
+  subscribe(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'SUBSCRIBE',
+      type: 'route',
+    })
+    return this
+  }
+  unsubscribe(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'UNSUBSCRIBE',
+      type: 'route',
+    })
+    return this
+  }
+  trace(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'TRACE',
+      type: 'route',
+    })
+    return this
+  }
+  acl(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'ACL',
+      type: 'route',
+    })
+    return this
+  }
+  connect(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'CONNECT',
+      type: 'route',
+    })
+    return this
+  }
+  bind(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'BIND',
+      type: 'route',
+    })
+    return this
+  }
+  unbind(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'UNBIND',
+      type: 'route',
+    })
+    return this
+  }
+  rebind(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'REBIND',
+      type: 'route',
+    })
+    return this
+  }
+  link(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'LINK',
+      type: 'route',
+    })
+    return this
+  }
+  unlink(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'UNLINK',
+      type: 'route',
+    })
+    return this
+  }
+  mkcalendar(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'MKCALENDAR',
+      type: 'route',
+    })
+    return this
+  }
+  propfind(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'PROPFIND',
+      type: 'route',
+    })
+    return this
+  }
+  proppatch(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'PROPPATCH',
+      type: 'route',
+    })
+    return this
+  }
+  source(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
+    pushMiddleware(this.middleware)({
+      path,
+      handler,
+      handlers,
+      method: 'SOURCE',
+      type: 'route',
+    })
+    return this
+  }
   all(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
     for (const method of METHODS) {
       pushMiddleware(this.middleware)({
@@ -163,6 +454,12 @@ export class Router {
     }
     return this
   }
+  /**
+   * Push middleware to the stack
+   * @param path path that middleware will handle if request URL starts with it
+   * @param handler handler function
+   * @param handlers the rest handler functions
+   */
   use(path: string | Handler, handler?: Handler, ...handlers: Handler[]) {
     pushMiddleware(this.middleware)({
       path,