+++ packages/etag/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tinyhttp/etag
 
+## 0.2.4
+
+### Patch Changes
+
+- Upgrade dependencies of all of the packages
+
 ## 0.2.3
 
 ### Patch Changes