+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @tinyhttp/logger
 
+## 0.3.8
+
+### Patch Changes
+
+- Update cookie-parser imports
+
+## 0.3.1
+
+### Patch Changes
+
+- Refactor logger
+
+## 0.3.0
+
+### Patch Changes
+
+- Added LoggerProperties interface for instantiating the logger
+- Added timestamp option with formatting.
+
+## 0.2.4
+
+### Patch Changes
+
+- Upgrade dependencies of all of the packages
+
 ## 0.2.3
 
 ### Patch Changes